Converting to work with .arclint files
Some changes were necessary to make this library work with the syntax
of the .arclint file and the ArcanistConfigurationDrivenLintEngine
engine.

This removes the need for the .arcconfig changes and allows
configuration through the .arclint file.

I didn’t try to make it work with both  .arcconfig and .arclint setups,
as we don’t use the .arcconfig style since it only allows one linter.